+++ forui/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.17.0 (Next)
 
+### `FTappable`
+*  Change `FTappable` to only update focused state when it has primary focus.
+
+
 ### Others
 * **Breaking** Remove `FSidebar.width`. This field was never intended for public usage - Use `FSidebarStyle.constraints`
   instead.